Refactor BranchInjector code to be in Branch module rather than view controller
